Rewrite set parser to be more flexible
The set parser can now handle set input without assingment. For starters
this enables the "element of" operator to work. This also adds a degree
of flexability for shorter lists.
