+++ src/util/parser/set/ast.ts
@@ -1,3 +1,21 @@
 import type { List, Set } from "immutable"
 
 export type SetElement = string | Set<SetElement> | List<string>
+
+export type Node = SetElement | SetExprNode
+
+export type SetExprNode =
+  | {
+      id: string
+      left: Node
+      right: Node
+    }
+  | string
+
+export function createNode(id: string, left: Node, right: Node): Node {
+  return {
+    id: id,
+    left: left,
+    right: right
+  }
+}
